refactor: useFunnel로 페이지 분리
