+++ src/pages/course/page.tsx
@@ -2,20 +2,17 @@ import { useGetAllCourses } from "../../react-query/queries/queries";
 import 'swiper/css';
 import 'swiper/css/effect-coverflow';
 import { sampleData } from "./test";
-import { useParams } from "react-router-dom";
-import DetailPage from "./DetailPage";
-import OverviewPage from "./OverViewPage";
-import ShareCorsePage from "./ShareCorsePage";
+import SharedCoursePage from "./SharedCourse/SharedCoursePage";
+import ShareCorsePage from "./ShareCourse/ShareCorsePage";
 
 export default function Page(){
-  const courseId = useParams().courseId;
 
   // 전체 일정 경로 조회
   const allCoursesQueries = useGetAllCourses(sampleData);
 
   const isLoading = allCoursesQueries.some(query => query.isLoading);
   const isError = allCoursesQueries.some(query => query.isError);
-  const isEnd = false;
+  const isEnd = true;
 
   if(isLoading) return <div style={{textAlign:"center",color:"white"}}>Loading...</div>;
   if(isError) return <div style={{textAlign:"center",color:"white"}}>Error...</div>;
@@ -28,21 +25,10 @@ export default function Page(){
     </>
   )
 
-  // courseId가 있으면 해당 코스의 상세 페이지를, 없으면 전체 개요 페이지를 렌더링
-  if (courseId) {
-    return (
-      <>
-        <DetailPage 
-          allCoursesQueries={allCoursesQueries}
-        />
-      </>
-    );
-  }
-
   return (
     <>
-      <OverviewPage   
-        dayOnCourseQueries={allCoursesQueries[0].data}
+      <SharedCoursePage   
+        allCoursesQueries={allCoursesQueries}
       />
     </>
   );
